# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Teslog API
  description: >
    Self-hosted Tesla vehicle data logging platform. Tracks drives, charges,
    idle sessions, battery health, and firmware updates via Tesla Fleet
    Telemetry.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: "{server_url}/api"
    variables:
      server_url:
        default: http://localhost
        description: Base URL of the Teslog instance
...
security:
  - bearerAuth: []
paths:
...
  /v1/vehicles:
    get:
      tags: [Vehicles]
      summary: List all vehicles
      operationId: listVehicles
      responses:
        "200":
          description: Vehicle list (includes latest state)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      tags: [Vehicles]
      summary: Create a vehicle manually
      operationId: createVehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 255
                vin:
                  type: string
                  maxLength: 17
                tesla_vehicle_id:
                  type: string
                model:
                  type: string
                trim:
                  type: string
                color:
                  type: string
      responses:
        "201":
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /v1/drives:
    get:
      tags: [Drives]
      summary: List drives (paginated)
      operationId: listDrives
      parameters:
        - name: vehicle_id
          in: query
          schema:
            type: integer
        - name: tag
          in: query
          schema:
            type: string
        - name: from
          in: query
          description: ISO 8601 datetime lower bound on started_at
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: ISO 8601 datetime upper bound on started_at
          schema:
            type: string
            format: date-time
        - $ref: "#/components/parameters/perPage"
        - $ref: "#/components/parameters/page"
      responses:
        "200":
          description: Paginated drive list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDrives"
  /v1/drives/{driveId}:
    get:
      tags: [Drives]
      summary: Get a single drive
      operationId: getDrive
      parameters:
        - $ref: "#/components/parameters/driveId"
      responses:
        "200":
          description: Drive with vehicle and place relations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drive"
    put:
      tags: [Drives]
      summary: Update drive tag or notes
      operationId: updateDrive
      parameters:
        - $ref: "#/components/parameters/driveId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tag:
                  type: string
                  maxLength: 100
                  nullable: true
                notes:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Updated drive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drive"
...
  /v1/charges:
    get:
      tags: [Charges]
      summary: List charges (paginated)
      operationId: listCharges
      parameters:
        - name: vehicle_id
          in: query
          schema:
            type: integer
        - name: charge_type
          in: query
          schema:
            type: string
            enum: [ac, dc, supercharger]
        - $ref: "#/components/parameters/perPage"
        - $ref: "#/components/parameters/page"
      responses:
        "200":
          description: Paginated charge list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCharges"
  /v1/charges/{chargeId}:
    get:
      tags: [Charges]
      summary: Get a single charge
      operationId: getCharge
      parameters:
        - $ref: "#/components/parameters/chargeId"
      responses:
        "200":
          description: Charge with vehicle and place relations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charge"
    put:
      tags: [Charges]
      summary: Update charge tag or notes
      operationId: updateCharge
      parameters:
        - $ref: "#/components/parameters/chargeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tag:
                  type: string
                  maxLength: 100
                  nullable: true
                notes:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Updated charge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charge"
...
  /v1/places:
    get:
      tags: [Places]
      summary: List all saved places
      operationId: listPlaces
      responses:
        "200":
          description: Place list with TOU rates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Place"
    post:
      tags: [Places]
      summary: Create a place
      operationId: createPlace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, latitude, longitude]
              properties:
                name:
                  type: string
                  maxLength: 255
                latitude:
                  type: number
                  format: double
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  format: double
                  minimum: -180
                  maximum: 180
                radius_meters:
                  type: integer
                  minimum: 10
                  maximum: 5000
                electricity_cost_per_kwh:
                  type: number
                  nullable: true
                auto_tag:
                  type: string
                  maxLength: 100
                  nullable: true
      responses:
        "201":
          description: Place created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /v1/export/drives:
    get:
      tags: [Export]
      summary: Export drives as CSV
      operationId: exportDrives
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="teslog-drives.csv"'
  /v1/export/charges:
    get:
      tags: [Export]
      summary: Export charges as CSV
      operationId: exportCharges
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="teslog-charges.csv"'

  # ── Commands ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Sanctum API token obtained from /v1/auth/login or /v1/auth/register
    telemetryAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Shared secret for fleet-telemetry ingestion
  parameters:
...
    driveId:
      name: driveId
      in: path
      required: true
      schema:
        type: integer
    chargeId:
      name: chargeId
      in: path
      required: true
      schema:
        type: integer
...
    perPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 25
        maximum: 100
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
  schemas:
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        tesla_vehicle_id:
          type: string
          nullable: true
        vin:
          type: string
          nullable: true
        name:
          type: string
        model:
          type: string
          nullable: true
        trim:
          type: string
          nullable: true
        battery_capacity_kwh:
          type: number
          nullable: true
        color:
          type: string
          nullable: true
        firmware_version:
          type: string
          nullable: true
        is_active:
          type: boolean
        show_on_dashboard:
          type: boolean
        tesla_token_expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        latest_state:
          $ref: "#/components/schemas/VehicleState"
    VehicleState:
      type: object
      properties:
        id:
          type: integer
        vehicle_id:
          type: integer
        timestamp:
          type: string
          format: date-time
        state:
          type: string
          description: "Vehicle state (e.g. driving, charging, idle, asleep)"
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        heading:
          type: number
          nullable: true
        elevation:
          type: number
          nullable: true
        speed:
          type: number
          nullable: true
        power:
          type: number
          nullable: true
        battery_level:
          type: number
          nullable: true
        rated_range:
          type: number
          nullable: true
        ideal_range:
          type: number
          nullable: true
        odometer:
          type: number
          nullable: true
        inside_temp:
          type: number
          nullable: true
        outside_temp:
          type: number
          nullable: true
        locked:
          type: boolean
          nullable: true
        sentry_mode:
          type: boolean
          nullable: true
        climate_on:
          type: boolean
          nullable: true
        gear:
          type: string
          nullable: true
        charger_power:
          type: number
          nullable: true
        charger_voltage:
          type: number
          nullable: true
        charger_current:
          type: number
          nullable: true
        charge_limit_soc:
          type: number
          nullable: true
        charge_state:
          type: string
          nullable: true
        energy_remaining:
          type: number
          nullable: true
        software_version:
          type: string
          nullable: true
    Drive:
      type: object
      properties:
        id:
          type: integer
        vehicle_id:
          type: integer
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
        distance:
          type: number
          nullable: true
          description: Distance in miles
        energy_used_kwh:
          type: number
          nullable: true
        efficiency:
          type: number
          nullable: true
          description: Wh/mi
        start_latitude:
          type: number
          format: double
          nullable: true
        start_longitude:
          type: number
          format: double
          nullable: true
        end_latitude:
          type: number
          format: double
          nullable: true
        end_longitude:
          type: number
          format: double
          nullable: true
        start_address:
          type: string
          nullable: true
        end_address:
          type: string
          nullable: true
        start_place_id:
          type: integer
          nullable: true
        end_place_id:
          type: integer
          nullable: true
        start_battery_level:
          type: number
          nullable: true
        end_battery_level:
          type: number
          nullable: true
        start_rated_range:
          type: number
          nullable: true
        end_rated_range:
          type: number
          nullable: true
        start_odometer:
          type: number
          nullable: true
        end_odometer:
          type: number
          nullable: true
        max_speed:
          type: number
          nullable: true
        avg_speed:
          type: number
          nullable: true
        outside_temp_avg:
          type: number
          nullable: true
        tag:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        weather:
          type: object
          nullable: true
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        start_place:
          $ref: "#/components/schemas/Place"
        end_place:
          $ref: "#/components/schemas/Place"
...
    Charge:
      type: object
      properties:
        id:
          type: integer
        vehicle_id:
          type: integer
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
        charge_type:
          type: string
          enum: [ac, dc, supercharger]
          nullable: true
        energy_added_kwh:
          type: number
          nullable: true
        energy_used_kwh:
          type: number
          nullable: true
        charging_efficiency:
          type: number
          nullable: true
        cost:
          type: number
          nullable: true
        start_battery_level:
          type: number
          nullable: true
        end_battery_level:
          type: number
          nullable: true
        start_rated_range:
          type: number
          nullable: true
        end_rated_range:
          type: number
          nullable: true
        max_charger_power:
          type: number
          nullable: true
        avg_voltage:
          type: number
          nullable: true
        max_voltage:
          type: number
          nullable: true
        avg_current:
          type: number
          nullable: true
        max_current:
          type: number
          nullable: true
        odometer:
          type: number
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        address:
          type: string
          nullable: true
        place_id:
          type: integer
          nullable: true
        tag:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        place:
          $ref: "#/components/schemas/Place"
...
    Place:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        radius_meters:
          type: integer
        electricity_cost_per_kwh:
          type: number
          nullable: true
        auto_tag:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        tou_rates:
          type: array
          items:
            $ref: "#/components/schemas/PlaceTouRate"
    PlaceTouRate:
      type: object
      properties:
        id:
          type: integer
        place_id:
          type: integer
        day_of_week:
          type: integer
          description: "0 = Sunday, 6 = Saturday"
        start_time:
          type: string
          example: "22:00"
        end_time:
          type: string
          example: "06:00"
        rate_per_kwh:
          type: number
...
    PaginatedDrives:
      type: object
      properties:
        current_page:
          type: integer
        data:
          type: array
          items:
            $ref: "#/components/schemas/Drive"
        first_page_url:
          type: string
        from:
          type: integer
          nullable: true
        last_page:
          type: integer
        last_page_url:
          type: string
        links:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                nullable: true
              label:
                type: string
              active:
                type: boolean
        next_page_url:
          type: string
          nullable: true
        path:
          type: string
        per_page:
          type: integer
        prev_page_url:
          type: string
          nullable: true
        to:
          type: integer
          nullable: true
        total:
          type: integer
    PaginatedCharges:
      type: object
      properties:
        current_page:
          type: integer
        data:
          type: array
          items:
            $ref: "#/components/schemas/Charge"
        first_page_url:
          type: string
        from:
          type: integer
          nullable: true
        last_page:
          type: integer
        last_page_url:
          type: string
        links:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                nullable: true
              label:
                type: string
              active:
                type: boolean
        next_page_url:
          type: string
          nullable: true
        path:
          type: string
        per_page:
          type: integer
        prev_page_url:
          type: string
          nullable: true
        to:
          type: integer
          nullable: true
        total:
          type: integer
...